--- a/datafusion/core/src/physical_optimizer/test_utils.rs
+++ b/datafusion/core/src/physical_optimizer/test_utils.rs
@@ -118,7 +118,12 @@ impl QueryCase {
         if error.is_some() {
             let plan_error = plan.unwrap_err();
             let initial = error.unwrap().to_string();
-            assert!(plan_error.to_string().contains(initial.as_str()));
+            assert!(
+                plan_error.to_string().contains(initial.as_str()),
+                "plan_error: {:?} doesn't contain message: {:?}",
+                plan_error,
+                initial.as_str()
+            );
         } else {
             assert!(plan.is_ok())
         }