@ozankabak 2023-01-26
I did an initial round of review on this, and subsequently made some simplifications and added comments. I have a few questions we should go over in our next meeting.

As I was changing the base branch (since the dependent API change branch merged upstream already), I unfortunately lost our commit history, so we will be starting from scratch (in terms of history) but that's OK.