@mingmwang 2023-01-30
I plan to do some refactoring work to the `EnforceSorting` rule. I think currently it was too complex to understand.