@ozankabak 2023-01-30
> I plan to do some refactoring work to the `EnforceSorting` rule. I think currently it was too complex to understand.

@mingmwang: Sounds good, as long as we don't lose functionality I am always on board with refactors. This PR has some important additions, let us finish this (a couple days at most) and add some more tests to avoid losing any functionality during refactors. With that, you can safely take a look at how you can simplify it and refactor.