@mustafasrepo 2023-01-31
> I can understand the final purpose of this PR. I will review this PR carefully tomorrow. One thing that I would like to hear from you guys is for the `SortExec`, should we add a flag to indicate it is local sort or global sort ? This will make the final physical plans easy to understand. And when we do Sort related optimizations, it is less error prone.
> 
> @mustafasrepo @ozankabak

Currently we check if `SortExec` preserves partitioning and its input has partitioning size 1; to decide for whether SortExec is global or local. I think using a flag would be useful.